Added XLog Decoding to jk_kafka_consumer to GPDB
The jk_kafka_consumer is now able to decode and parse incoming messages.
If the expected segment number is given with the `-s` option, the
consumer will assume the incoming data begins with offset 0 of that
segment file. The decoding logic (using xlogreader) is similar to that
of xlogdump, where the rmgrdesc is used to get XLOG types.