Add hacked-up consistency point xlog record
A new function gp_create_cluster_consistent_replay_point(TEXT) was added
to the catalog. It blocks all transactions and creates an XLOG of type
XLOG_CONSISTENCY_POINT that gets sent out to the master and every
segment. According rmgr and desc info added.

The function is hacked-up in the sense that it takes the 2pc lock that
holds more power than this function needs.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Junkeun Yi <juyi@pivotal.io>