Make the kafka_producer to use topic basedir's basename
Previously, the kafka producer that was embedded into the receivelog.c
code was strictly set to 'gptest'. Now, it will be the basename of the
base directory given to receivelog.
E.g. If the receivelog was invoked by `pg_receivexlog -D /tmp/primary1`,
the kafka topic will be set to `primary1`.

Co-authored-by: Shivram Mani <smani@pivotal.io>
Co-authored-by: Junkeun Yi <juyi@pivotal.io>